pyproject.toml
[project]
name = "mcp-gateway"
version = "0.1.0"
description = "Modular MCP gateway for web search, fetch, browse, and document parsing"
requires-python = ">=3.11"
dependencies = [
    # Core MCP
    "mcp[cli]>=1.0.0",

    # HTTP client
    "httpx>=0.27.0",

    # Web content extraction
    "markdownify>=0.13.0",
    "trafilatura>=2.0.0",
    "readability-lxml>=0.8.1",

    # Server
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",

    # Document parsing
    "docling>=2.0.0",
]

[project.optional-dependencies]
# For GPU acceleration (optional)
gpu = [
    "torch>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[project.scripts]
mcp-gateway = "gateway:run_stdio"
mcp-gateway-sse = "gateway:run_sse"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
